// src/app/portal/moduls/customer-service/components/options-event-types/models/eventTypes.ts
export class EventTypesClass {
    EventType: string;
    SubjectOfTheEvent: string;
    ApplicantDocumentType: string | null;
    ApplicantDocument: string | null;
    ThirdDocumentType: string;
    ThirdDocument: string;
    RelatedArea: string;
    NamesOfficial: string;
    EventDate: string;
    EventCodeDepartment: string;
    EventCodeCity: string;
    EventComment: string;
    AcceptTermsAndConditions: boolean;
    SendFileThanksAndCongratulations: any;
    SendFilesComplaintsAndClaims: any;
    SendBeneficiaryDeathCertificate: any;
    SendFileBankAccount: any;
    SendFileTitularDeathCertificate: any;
    SendFilesCivilRegistrationMarriage: any;
    SendFileBirthCertificateSon: any;
    SendFileSpouseDeathCertificate: any;
    SendFileExtraJudgmentStatement: any;
    SendFileBankAccountCertificate: any;
    SendFileBankStatementOrRemovablePayroll: any;
    ReasonsForReimbursement: string;
    BankAccountUser: boolean;
}